backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiohttp==3.9.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pandas==2.1.4
numpy==1.25.2
python-multipart==0.0.6
pydantic==2.5.0
redis==5.0.1
celery==5.3.4
